--- user-interface/src/main/java/life/qbic/datamanager/views/general/download/DownloadContentProvider.java
@@ -1,10 +1,0 @@
-package life.qbic.datamanager.views.general.download;
-
-/**
- * Provides content and file name for any files created from data and metadata.
- */
-public interface DownloadContentProvider {
-
-  public byte[] getContent();
-  public String getFileName();
-}
