+++ user-interface/src/main/java/life/qbic/datamanager/views/projects/project/experiments/experiment/ExperimentDetailsComponent.java
@@ -21,12 +21,7 @@ import com.vaadin.flow.router.RouteParameters;
 import com.vaadin.flow.server.StreamResource;
 import com.vaadin.flow.spring.annotation.SpringComponent;
 import com.vaadin.flow.spring.annotation.UIScope;
-import com.vaadin.flow.theme.lumo.LumoIcon;
-import java.io.BufferedWriter;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.OutputStream;
-import java.io.OutputStreamWriter;
 import java.io.Serial;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -40,6 +35,8 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import life.qbic.application.commons.ApplicationException;
+import life.qbic.datamanager.download.DownloadContentProvider.XLSXDownloadContentProvider;
+import life.qbic.datamanager.download.DownloadProvider;
 import life.qbic.datamanager.templates.TemplateService;
 import life.qbic.datamanager.views.Context;
 import life.qbic.datamanager.views.general.ConfirmEvent;
@@ -47,8 +44,6 @@ import life.qbic.datamanager.views.general.Disclaimer;
 import life.qbic.datamanager.views.general.PageArea;
 import life.qbic.datamanager.views.notifications.CancelConfirmationDialogFactory;
 import life.qbic.datamanager.views.notifications.MessageSourceNotificationFactory;
-import life.qbic.datamanager.views.general.download.DownloadContentProvider;
-import life.qbic.datamanager.views.general.download.DownloadProvider;
 import life.qbic.datamanager.views.projects.project.experiments.ExperimentInformationMain;
 import life.qbic.datamanager.views.projects.project.experiments.experiment.components.CardCollection;
 import life.qbic.datamanager.views.projects.project.experiments.experiment.components.ExistingGroupsPreventVariableEdit;
@@ -77,9 +72,7 @@ import life.qbic.projectmanagement.domain.model.experiment.VariableLevel;
 import life.qbic.projectmanagement.domain.model.project.Project;
 import life.qbic.projectmanagement.domain.model.project.ProjectId;
 import life.qbic.projectmanagement.domain.model.sample.Sample;
-import life.qbic.projectmanagement.domain.model.sample.SampleOrigin;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
-import org.checkerframework.checker.units.qual.A;
 import org.springframework.beans.factory.annotation.Autowired;
 
 /**
@@ -146,23 +139,11 @@ public class ExperimentDetailsComponent extends PageArea {
     this.addClassName("experiment-details-component");
     Button download = new Button("Registration Template");
     download.addClickListener(buttonClickEvent -> {
-      try (XSSFWorkbook workbook =templateService.sampleBatchRegistrationTemplate(context.projectId().orElseThrow().value(), context.experimentId().orElseThrow().value())) {
-        DownloadProvider downloadProvider = new DownloadProvider(new DownloadContentProvider() {
-          @Override
-          public byte[] getContent() {
-            try (ByteArrayOutputStream arrayOutputStream = new ByteArrayOutputStream()) {
-              workbook.write(arrayOutputStream);
-              return arrayOutputStream.toByteArray();
-            } catch (IOException e) {
-              throw new RuntimeException(e);
-            }
-          }
-
-          @Override
-          public String getFileName() {
-            return context.projectId().orElseThrow().value() + "_registration_template.xlsx";
-          }
-        });
+      try (XSSFWorkbook workbook = templateService.sampleBatchRegistrationXLSXTemplate(
+          context.projectId().orElseThrow().value(),
+          context.experimentId().orElseThrow().value())) {
+        var downloadProvider = new DownloadProvider(new XLSXDownloadContentProvider(
+            context.projectId().orElseThrow().value() + "_registration_template.xlsx", workbook));
         add(downloadProvider);
         downloadProvider.trigger();
       } catch (IOException e) {
@@ -172,23 +153,11 @@ public class ExperimentDetailsComponent extends PageArea {
     this.add(download);
     Button updateTemplate = new Button("Update Template");
     updateTemplate.addClickListener(buttonClickEvent -> {
-      try (XSSFWorkbook workbook =templateService.sampleBatchUpdateTemplate(context.projectId().orElseThrow().value(), context.experimentId().orElseThrow().value())) {
-        DownloadProvider downloadProvider = new DownloadProvider(new DownloadContentProvider() {
-          @Override
-          public byte[] getContent() {
-            try (ByteArrayOutputStream arrayOutputStream = new ByteArrayOutputStream()) {
-              workbook.write(arrayOutputStream);
-              return arrayOutputStream.toByteArray();
-            } catch (IOException e) {
-              throw new RuntimeException(e);
-            }
-          }
-
-          @Override
-          public String getFileName() {
-            return context.projectId().orElseThrow().value() + "_update_template.xlsx";
-          }
-        });
+      try (XSSFWorkbook workbook = templateService.sampleBatchUpdateXLSXTemplate(
+          context.projectId().orElseThrow().value(),
+          context.experimentId().orElseThrow().value())) {
+        DownloadProvider downloadProvider = new DownloadProvider(new XLSXDownloadContentProvider(
+            context.projectId().orElseThrow().value() + "_update_template.xlsx", workbook));
         add(downloadProvider);
         downloadProvider.trigger();
       } catch (IOException e) {
