+++ user-interface/src/main/java/life/qbic/datamanager/views/projects/project/samples/SampleInformationMain.java
@@ -21,12 +21,12 @@ import java.util.Collection;
 import java.util.Comparator;
 import java.util.List;
 import life.qbic.application.commons.ApplicationException;
+import life.qbic.datamanager.download.DownloadProvider;
 import life.qbic.datamanager.views.AppRoutes.Projects;
 import life.qbic.datamanager.views.Context;
 import life.qbic.datamanager.views.general.Disclaimer;
 import life.qbic.datamanager.views.general.DisclaimerConfirmedEvent;
 import life.qbic.datamanager.views.general.Main;
-import life.qbic.datamanager.views.general.download.DownloadProvider;
 import life.qbic.datamanager.views.notifications.CancelConfirmationDialogFactory;
 import life.qbic.datamanager.views.notifications.MessageSourceNotificationFactory;
 import life.qbic.datamanager.views.notifications.StyledNotification;
